WIP: save - refactor dashboard, extract hook, constants
